reduce size of QrDetailActiviry.java
